Revert back for testing

+++ pipelines/skylab/snm3C/snm3C.wdl
@@ -221,11 +221,8 @@ task Demultiplexing {
             adapter_name = re.search(r'A(\d+)-R', filename)
             if adapter_name:
                 adapter_name = 'A' + adapter_name.group(1)
-                if adapter_name in adapter_counts:
-                    if (adapter_counts[adapter_name] < min_threshold) or (adapter_counts[adapter_name] > max_threshold):
-                        print("Removing file: ", file_path) 
-                        print(adapter_name, adapter_counts[adapter_name])
-                        os.remove(file_path)
+                if adapter_name in adapter_counts and adapter_counts[adapter_name] > max_threshold:
+                    os.remove(file_path)
     CODE
 
     # Batch the fastq files into folders of batch_number size
